Change from Resource to Prompts
