+++ src/Mcp/Prompts/LaravelCodeSimplifier/laravel-code-simplifier.blade.php
