--- src/Mcp/Resources/LaravelCodeSimplifier/LaravelCodeSimplifier.php
@@ -1,27 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Laravel\Boost\Mcp\Resources\LaravelCodeSimplifier;
-
-use Laravel\Boost\Mcp\Prompts\Concerns\RendersBladeGuidelines;
-use Laravel\Mcp\Response;
-use Laravel\Mcp\Server\Resource;
-
-class LaravelCodeSimplifier extends Resource
-{
-    use RendersBladeGuidelines;
-
-    protected string $description = 'Simplifies and refines PHP/Laravel code for clarity, consistency, and maintainability while preserving all functionality. Focuses on recently modified code unless instructed otherwise.';
-
-    protected string $uri = 'file://instructions/laravel-code-simplifier.md';
-
-    protected string $mimeType = 'text/markdown';
-
-    public function handle(): Response
-    {
-        $content = $this->renderGuidelineFile(__DIR__.'/laravel-code-simplifier.blade.php');
-
-        return Response::text($content);
-    }
-}
